extract main API from Assert
This extracts the stubbing API from Assert and reworks it as
MuchStub.  All of the API is remaining the same, but instead of
calling `Assert.{stub|unstub|stub_send|etc}` you now call
`MuchStub.{stub|unstub|stub_send|etc}`.  The usage doesn't change.

Note: I chose to move the custom errors classes and the `NullStub`
and `ParameterList` under the main `MuchStub` namespace.  Assert
had them under the `Assert::Stub` namespace, but I'm not sure why.
It makes more sense to have them just below the main `MuchStub`
namespace.

Note: the only other difference is that Assert auto-unstubs on
test context teardowns.  It isn't appropriate and is not the
intent of MuchStub to auto do this (b/c it would differ based on
which testing framework you used).  Therefore, the MuchStub test
helpers manually add a teardown to unstub their Assert test
contexts.

Note: this also pulls over the `caller_locations` Ruby 1.8.7
polyfill from Assert.  This is needed for 1.8.7 compatibility.
We prefer `caller_locations` as it is more performant than
`caller`.