add `MuchStub.call` for creating stubs; allows `MuchStub.()` API
This is a more terse and UX-friendly way to build a stub and
emphasizes how creating stubs is the main point of MuchStub.
Plus, the `MuchStub.stub` API is kinda redundant.

I chose to keep the `MuchStub.stub` API for backwards compatibility
reasons and b/c it doesn't hurt anything if users prefer using it.