update README
This polishes the README by removing the abbreviations. In watching
other developers consume the README, it became apparent tha the
abbreviations were a bit of a distraction. This removes them to
avoid distracting users and make learning about MuchStub as easier.

This also adds a couple of sections for using generic stub blocks
to spy on method calls. This adds various examples of basic
argument spying and also test double spying.

All of this is to ease new developers in learning how to use
MuchStub and what types of tests are possible using it.