add `MuchStub.tap`
This method is used to "tap" using a stub. This allows you to spy
on a method call while preserving the methods original behavior and
return value. You can also use this to late-bound stub the return
value of a stubbed method.

This is extracted from use-patterns we've seen using MuchStub in
production apps.