Fix bug when building CallSpy with `nil` return values
Before this change, when a CallSpy was created with `nil` return
values, those values would never get returned. This was because
`call_spy_return_value_proc` method was always returning the
instance of the CallSpy when the return value was `nil`.

That method was attempting to detect whether a return value had
been specified, however, a `nil` check is not sufficient because
a user could intend to specify a `nil` return value. A better
approach for this is to do a `key` check on the return values
hash. I chose to use `fetch` with a default value to avoid an
extra conditional.